backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9      # PostgreSQL driver
alembic==1.13.3             # Database migrations

# Validation & settings
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9     # required for OAuth2 form parsing

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.3.2
httpx==0.27.2               # async test client for FastAPI
pytest-asyncio==0.23.8
